+++ tapeserver/castor/tape/tapeserver/daemon/Payload.hpp
@@ -41,13 +41,12 @@ class Payload
   Payload(const Payload&);
   Payload& operator=(const Payload&);
 public:
-  Payload(size_t capacity):
-  m_data(new (std::nothrow) unsigned char[capacity]),m_totalCapacity(capacity),m_size(0) {
-    if(nullptr == m_data) {
+  Payload(uint32_t capacity) : m_data(new(std::nothrow) unsigned char[capacity]), m_totalCapacity(capacity), m_size(0) {
+    if (nullptr == m_data) {
       throw cta::exception::MemException("Failed to allocate memory for a new MemBlock!");
     }
   }
-  
+
   ~Payload(){
     delete[] m_data;
   }
